Reordering unit tests in single files per each
